requirements.txt
# Face recognition (dlib-based, produces 128-D embeddings)
face-recognition>=1.3.0

# OCR
pytesseract>=0.3.10

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Clustering
scikit-learn>=1.3.0

# CLI
click>=8.1.0
rich>=13.0.0

# Progress bars
tqdm>=4.65.0

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
